Use methods deprecated by Spec2
Although the comment indicates intent to deprecate, there is no apparent working replacement for SpPresenter on:  when instantiating standalone presenters.  This commit removes the "fix" in the previous UI commit for Logic-Boolean, because it broke the demos.   Note that this returns Spec to its previous behavior: to change a SpPresenterWithModel's model after initialization, one must  break encapsulation by sending modelChanged manually.

BLFunctionPresenter>>title returns a String rather than a presentation object, on the assumption that since Spec2 is deprecating the latter use,  the message is available for our use.